rename utils module in login app
